+++ manejo_ocr.py
